Add support for struct pointer traversal
This deals with two cases: where we have been provided with a pointer to
a struct in the call to Marshal(), and also where we encounter a pointer
while encoding a struct. The former case resulted in a panic (see #13),
while the latter reported that Pointer types were not supported. The
solution uses reflection to continually walk across pointers to get to
the underlying value.

Fixes #13