Add support for documenting TOML fields
This adds the ability to have fields documented with comments in
marshalled TOML files. This is facilitated through the use of a "doc"
struct tag. Any string appearing in the "doc" struct tag is appended to
the field in the marshalled TOML as a TOML comment. For structs,
comments precede the generated TOML table.