Add MarshalerRec, UnmarshalerRec and improve documentation
Motivation for the new interfaces: There is often a need to validate
values after unmarshaling them. In other cases, the desired TOML
encoding might not match fields of a Go struct value exactly. The
standard way to handle such needs with encoding/json is the 'recursive
decoding' pattern:

    func (x *Type) UnmarshalJSON(b []byte) error {
        var enc otherType
        if err := json.Unmarshal(b); err != nil {
             return err
        }
        // more validation here, then assign to *x.
    }

This pattern can't be used with the existing Unmarshaler interface
because it deals with raw TOML table syntax.

* The syntax of a table always includes the full key path. Recursive
  decoding can't work because the key path depends on the type in which
  the Unmarshaler is contained.
* Decoding a value needs to know about TOML syntax. package toml cannot
  be used to unmarshal into a value (input must always be a table).

The new interfaces are inspired by gopkg.in/yaml.v2.

Fixes #22 because documentation is now backed by a GoDoc example.
Fixes #27
Fixes #34