Simplify field matching logic
findField, which runs for each key in the input, looked for a matching
named defined through the struct tag and then tried to come up with a
struct field name by transforming the key name in various ways.

For the example in #28, the key is 'api_key'. and findField tried
'Api_key', 'ApiKey', and 'API_KEY'. None of these match the field
'APIKey', which is created by the encoder using ToSnakeCase and its
built-in abbreviation table.

The new code has linear complexity and ensures that key names defined
through a struct tag are enforced. It was previously possible to set
such fields through their auto-mapped name. This is also prep-work for
user-defined field mapping functions.

For #28, the code now creates a cache containing 'apikey' as the mapped
name for the 'APIKey' field. The input key 'api_key' is also mapped to
'apikey' by the same function.

Fixes #28